qa: align Playwright specs with restructured QA catalog
- CHN-005 → CHN-003 (channel rename updates sidebar)
- MSG-012 → MSG-006 (clickable mention opens agent profile)
- ATT-001 merged into MSG-005 (attachment upload and download)
- MSG-002 gains Supersedes: MSG-002, DM-002
- Skip stubs for retired specs: AGT-004, PRF-001, DM-001, DM-002,
  KIMI-001, UNR-001, UNR-002
